refactor: show cursor shape glyphs in segmented control and move to Terminal (#145)
